Separate pixel and vector fonts into modules of their own
