+++ internal/core/software_renderer/fonts.rs
@@ -9,9 +9,9 @@ use core::cell::RefCell;
 use crate::thread_local_ as thread_local;
 
 use super::{PhysicalLength, PhysicalSize};
-use crate::graphics::{BitmapFont, BitmapGlyphs, FontRequest};
+use crate::graphics::{BitmapFont, FontRequest};
 use crate::lengths::{LogicalLength, LogicalSize, ScaleFactor};
-use crate::textlayout::{Glyph, TextLayout, TextShaper};
+use crate::textlayout::TextLayout;
 use crate::Coord;
 
 thread_local! {
@@ -42,131 +42,20 @@ pub trait GlyphRenderer {
     fn render_glyph(&self, glyph_id: core::num::NonZeroU16) -> RenderableGlyph;
 }
 
-impl BitmapGlyphs {
-    fn ascent(&self, font: &BitmapFont) -> PhysicalLength {
-        (PhysicalLength::new(self.pixel_size).cast() * font.ascent / font.units_per_em).cast()
-    }
-    fn descent(&self, font: &BitmapFont) -> PhysicalLength {
-        (PhysicalLength::new(self.pixel_size).cast() * font.descent / font.units_per_em).cast()
-    }
-    fn height(&self, font: &BitmapFont) -> PhysicalLength {
-        // The descent is negative (relative to the baseline)
-        (PhysicalLength::new(self.pixel_size).cast() * (font.ascent - font.descent)
-            / font.units_per_em)
-            .cast()
-    }
-    fn pixel_size(&self) -> PhysicalLength {
-        PhysicalLength::new(self.pixel_size)
-    }
-}
-
 pub(super) const DEFAULT_FONT_SIZE: LogicalLength = LogicalLength::new(12 as Coord);
 
-// A font that is resolved to a specific pixel size.
-pub struct PixelFont {
-    bitmap_font: &'static BitmapFont,
-    pub glyphs: &'static BitmapGlyphs,
-}
-
-impl PixelFont {
-    pub fn pixel_size(&self) -> PhysicalLength {
-        self.glyphs.pixel_size()
-    }
-    pub fn glyph_index_to_glyph_id(index: usize) -> core::num::NonZeroU16 {
-        core::num::NonZeroU16::new(index as u16 + 1).unwrap()
-    }
-    pub fn glyph_id_to_glyph_index(id: core::num::NonZeroU16) -> usize {
-        id.get() as usize - 1
-    }
-}
-
-impl GlyphRenderer for PixelFont {
-    fn render_glyph(&self, glyph_id: core::num::NonZeroU16) -> RenderableGlyph {
-        let glyph_index = Self::glyph_id_to_glyph_index(glyph_id);
-        let bitmap_glyph = &self.glyphs.glyph_data[glyph_index];
-        RenderableGlyph {
-            x: PhysicalLength::new(bitmap_glyph.x),
-            y: PhysicalLength::new(bitmap_glyph.y),
-            width: PhysicalLength::new(bitmap_glyph.width),
-            height: PhysicalLength::new(bitmap_glyph.height),
-            alpha_map: bitmap_glyph.data.as_slice().into(),
-        }
-    }
-}
-
-impl TextShaper for PixelFont {
-    type LengthPrimitive = i16;
-    type Length = PhysicalLength;
-    fn shape_text<GlyphStorage: core::iter::Extend<Glyph<PhysicalLength>>>(
-        &self,
-        text: &str,
-        glyphs: &mut GlyphStorage,
-    ) {
-        let glyphs_iter = text.char_indices().map(|(byte_offset, char)| {
-            let glyph_index = self
-                .bitmap_font
-                .character_map
-                .binary_search_by_key(&char, |char_map_entry| char_map_entry.code_point)
-                .ok()
-                .map(|char_map_index| {
-                    self.bitmap_font.character_map[char_map_index].glyph_index as usize
-                });
-            let x_advance = glyph_index.map_or_else(
-                || self.pixel_size(),
-                |glyph_index| PhysicalLength::new(self.glyphs.glyph_data[glyph_index].x_advance),
-            );
-            Glyph {
-                glyph_id: glyph_index.map(|index| Self::glyph_index_to_glyph_id(index)),
-                advance: x_advance,
-                text_byte_offset: byte_offset,
-                ..Default::default()
-            }
-        });
-        glyphs.extend(glyphs_iter);
-    }
-
-    fn glyph_for_char(&self, ch: char) -> Option<Glyph<PhysicalLength>> {
-        self.bitmap_font
-            .character_map
-            .binary_search_by_key(&ch, |char_map_entry| char_map_entry.code_point)
-            .ok()
-            .map(|char_map_index| {
-                let glyph_index =
-                    self.bitmap_font.character_map[char_map_index].glyph_index as usize;
-                let bitmap_glyph = &self.glyphs.glyph_data[glyph_index as usize];
-                let x_advance = PhysicalLength::new(bitmap_glyph.x_advance);
-                Glyph {
-                    glyph_id: Some(Self::glyph_index_to_glyph_id(glyph_index)),
-                    advance: x_advance,
-                    text_byte_offset: 0,
-                    ..Default::default()
-                }
-            })
-    }
-}
-
-impl crate::textlayout::FontMetrics<PhysicalLength> for PixelFont {
-    fn ascent(&self) -> PhysicalLength {
-        self.glyphs.ascent(self.bitmap_font)
-    }
-
-    fn descent(&self) -> PhysicalLength {
-        self.glyphs.descent(self.bitmap_font)
-    }
-
-    fn height(&self) -> PhysicalLength {
-        self.glyphs.height(self.bitmap_font)
-    }
-}
+mod pixelfont;
+#[cfg(feature = "systemfonts")]
+pub mod vectorfont;
 
 #[cfg(feature = "systemfonts")]
-pub(super) mod systemfonts;
+pub mod systemfonts;
 
 #[derive(derive_more::From)]
 pub enum Font {
-    PixelFont(PixelFont),
+    PixelFont(pixelfont::PixelFont),
     #[cfg(feature = "systemfonts")]
-    VectorFont(systemfonts::VectorFont),
+    VectorFont(vectorfont::VectorFont),
 }
 
 pub fn match_font(request: &FontRequest, scale_factor: ScaleFactor) -> Font {
@@ -211,7 +100,7 @@ pub fn match_font(request: &FontRequest, scale_factor: ScaleFactor) -> Font {
 
     let matching_glyphs = &font.glyphs[nearest_pixel_size];
 
-    PixelFont { bitmap_font: font, glyphs: matching_glyphs }.into()
+    pixelfont::PixelFont { bitmap_font: font, glyphs: matching_glyphs }.into()
 }
 
 pub fn text_layout_for_font<'a, Font: crate::textlayout::AbstractFont>(
